requirements.txt
# ── Core framework ──────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── LangChain ecosystem ───────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-mistralai>=0.1.0
langchain-huggingface>=0.0.3

# ── LLM ───────────────────────────────────────────────────────────────────────
mistralai>=1.0.0

# ── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers>=3.0.0
torch>=2.2.0          # CPU-only; swap for torch+cu121 if you have a GPU

# ── Vector databases ──────────────────────────────────────────────────────────
faiss-cpu>=1.8.0      # use faiss-gpu on CUDA machines
chromadb>=0.5.0       # optional Chroma backend

# ── PDF parsing ───────────────────────────────────────────────────────────────
pymupdf>=1.24.0       # primary parser  (imports as `fitz`)
pypdf>=4.2.0          # fallback parser

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
numpy>=1.26.0
